refactor_metrics2_cron_job
