+++ rest-api/singletons.py
@@ -14,6 +14,7 @@ GENERIC_SQL_DATABASE_INDEX = 5
 MAIN_CONFIG_INDEX = 6
 DB_CONFIG_INDEX = 7
 BACKUP_SQL_DATABASE_INDEX = 8
+ALEMBIC_SQL_DATABASE_INDEX = 9
 
 def reset_for_tests():
   with singletons_lock:
